feat: Added basic endpoint logic
